refactor: 💡 separate store (#82)
